+++ integration/taulgt.ps1
@@ -24,76 +24,8 @@
 
 #Requires -Version 7.3
 
-function Get-Logtalkhome {
-	if ($null -eq $env:LOGTALKHOME) 
-	{
-		Write-Output "The environment variable LOGTALKHOME should be defined first,"
-		Write-Output "pointing to your Logtalk installation directory!"
-		Write-Output "Trying the default locations for the Logtalk installation..."
-
-		$DEFAULTPATHS = [string[]](
-			"C:\Program Files (x86)\Logtalk",
-			"C:\Program Files\Logtalk",
-			"%LOCALAPPDATA%\Logtalk"
-		)
-
-		# Checking all default paths
-		foreach ($DEFAULTPATH in $DEFAULTPATHS) { 
-			Write-Output ("Looking for: " + $DEFAULTPATH)
-			if (Test-Path $DEFAULTPATH) {
-				Write-Output ("... using Logtalk installation found at " + $DEFAULTPATH)
-				$env:LOGTALKHOME = $DEFAULTPATH
-				break
-			}
-		}
-	}
-	# At the end LOGTALKHOME was set already or now is set
-}
-
-function Get-Logtalkuser {
-	if ($null -eq $env:LOGTALKUSER) {
-		Write-Output "After the script completion, you must set the environment variable"
-		Write-Output "LOGTALKUSER pointing to %USERPROFILE%\Documents\Logtalk."
-		$env:LOGTALKUSER = "%USERPROFILE%\Documents\Logtalk"
-	}
-	# At the end LOGTALKUSER was set already or now is set
-}
-
-Get-Logtalkhome
-
-# Check for existence
-if (!(Test-Path $env:LOGTALKHOME)) {
-	Write-Output "... unable to locate Logtalk installation directory!"
-	Write-Output ""
-	Start-Sleep -Seconds 2
-	Exit 1
-}
-
-Get-Logtalkuser
-
-# Check for existence
-if (Test-Path $env:LOGTALKUSER) {
-	if (!(Test-Path $env:LOGTALKUSER/VERSION.txt)) {
-		Write-Output "Cannot find VERSION.txt in the Logtalk user directory at %LOGTALKUSER%!"
-		Write-Output "Creating an up-to-date Logtalk user directory..."
-		logtalk_user_setup
-	} else {
-		$system_version = Get-Content $env:LOGTALKHOME/VERSION.txt
-		$user_version = Get-Content $env:LOGTALKUSER/VERSION.txt
-		if ($user_version -lt $system_version) {
-			Write-Output "Logtalk user directory at %LOGTALKUSER% is outdated: "
-			Write-Output "    $user_version < $system_version"
-			Write-Output "Creating an up-to-date Logtalk user directory..."
-			logtalk_user_setup
-		}
-	}
-} else {
-	Write-Output "Cannot find the Logtalk user directory at %LOGTALKUSER%!"
-	Write-Output "Running the logtalk_user_setup shell script to create the directory:"
-	logtalk_user_setup
-}
-
-$env:LOGTALK_STARTUP_DIRECTORY= $pwd
+Import-Module (Join-Path $PSScriptRoot "LogtalkSetup.psm1")
+Initialize-LogtalkEnvironment
 
 $source = $env:LOGTALKHOME + '\integration\logtalk_tau.js'