+++ RELEASE_NOTES.md
@@ -31,6 +31,9 @@ Prolog adapter and integration files
 * CHANGED: Moved all common setup code from the Bash shell integration scripts
 to a shared setup script.
 
+* CHANGED: Moved all common setup code from the PowerShell shell integration
+scripts to a module.
+
 * FIXED: The PowerShell integration scripts were not setting the `LOGTALKUSER`
 environment variable correctly when not defined by the user.